% Auto-generated: do not edit by hand
\name{''DashReactFlow}

\alias{''DashReactFlow}

\title{DashReactFlow component}

\description{

}

\usage{
''DashReactFlow(id=NULL, NodeClicked=NULL, attributionPosition=NULL,
className=NULL, defaultPosition=NULL, defaultZoom=NULL,
edges=NULL, fitView=NULL, maxZoom=NULL, minZoom=NULL,
nodes=NULL, nodesDraggable=NULL,
onlyRenderVisibleElements=NULL, preventScrolling=NULL,
snapGrid=NULL, snapToGrid=NULL, style=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{NodeClicked}{Named list. called when user clicks a node}

\item{attributionPosition}{Character. }

\item{className}{Character. }

\item{defaultPosition}{List of numerics. }

\item{defaultZoom}{Numeric. }

\item{edges}{List of named lists. array of edges (for an controlled flow)}

\item{fitView}{Logical. }

\item{maxZoom}{Numeric. }

\item{minZoom}{Numeric. }

\item{nodes}{List of named lists. array of nodes (for an controlled flow)}

\item{nodesDraggable}{Logical. }

\item{onlyRenderVisibleElements}{Logical. }

\item{preventScrolling}{Logical. }

\item{snapGrid}{List of numerics. }

\item{snapToGrid}{Logical. }

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

