% Auto-generated: do not edit by hand
\name{''DashReactFlow}

\alias{''DashReactFlow}

\title{DashReactFlow component}

\description{

}

\usage{
''DashReactFlow(id=NULL, NodeClicked=NULL, draggable=NULL, edges=NULL,
nodes=NULL, style=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{NodeClicked}{Named list. called when user clicks a node}

\item{draggable}{Logical. are the node and its edges draggable}

\item{edges}{List of named lists. array of edges (for an controlled flow)}

\item{nodes}{List of named lists. array of nodes (for an controlled flow)}

\item{style}{Named list. style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

